% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{basis_vector}
\alias{basis_vector}
\title{Generate basis vector in direction i in n dimensions (i <= n)}
\usage{
basis_vector(i, n)
}
\arguments{
\item{i}{selected direction}

\item{n}{number of dimensions}
}
\value{
basis vector
}
\description{
Generate basis vector in direction i in n dimensions (i <= n)
}
