% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{compareSmoothing}
\alias{compareSmoothing}
\title{Compare traces with different smoothing options.}
\usage{
compareSmoothing(d, tPath, idx, alphaV = c(0.01, 0.05, 0.1), n = 10)
}
\arguments{
\item{d}{Data matrix}

\item{tPath}{Interpolated tour path (as list of projections)}

\item{idx}{Index function}

\item{alphaV}{Jitter amounts to compare (for jittering angle or points)}

\item{n}{Number of evaluations entering mean value calculation}
}
\value{
Table of mean index values
}
\description{
Compare traces with different smoothing options.
}
\examples{
\donttest{
d <- spiralData(4, 100)
tPath <- tourr::save_history(d, max_bases=2)
tPath <- as.list(tourr::interpolate(tPath))
idx <- scagIndex("Skinny")
compS <- compareSmoothing(d, tPath, idx, n=5)
plotSmoothingComparison(compS)
}
}
