% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{get_index_mean}
\alias{get_index_mean}
\title{Evaluate mean index value over n jittered views.}
\usage{
get_index_mean(proj, d, alpha, idx, method = "jitter_angle", n = 10)
}
\arguments{
\item{proj}{Original projection plane}

\item{d}{Data matrix}

\item{alpha}{Jitter amount (for jittering angle or points)}

\item{idx}{Index function}

\item{method}{Select between "jitterAngle" (default) and "jitterPoints"
(otherwise we return original index value)}

\item{n}{Number of evaluations entering mean value calculation}
}
\value{
Mean index value
}
\description{
Evaluate mean index value over n jittered views.
}
