% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing.R
\name{getTrace}
\alias{getTrace}
\title{Tracing the index over interpolated planned tour path.}
\usage{
getTrace(d, m, indexList, indexLabels)
}
\arguments{
\item{d}{data}

\item{m}{list of projection matrices for the planned tour}

\item{indexList}{list of index functions to calculate for each entry}

\item{indexLabels}{labels used in the output}
}
\value{
index values for each interpolation step
}
\description{
Over an interpolated tour path between the planes specified
by basis vectors listed in x, y we calculate the index value
obtained for the data d in each projection. All index functions
must take the data in 2-d matrix format and return the index value.
}
