% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{jitter_points}
\alias{jitter_points}
\title{Re-evaluate index after jittering all points by an amount alpha.}
\usage{
jitter_points(proj_data, alpha, idx)
}
\arguments{
\item{proj_data}{Original projected data points}

\item{alpha}{Jitter amount (passed into the jitter() function)}

\item{idx}{Index function}
}
\value{
New index value
}
\description{
Re-evaluate index after jittering all points by an amount alpha.
}
