% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{jitterAngle}
\alias{jitterAngle}
\title{Re-evaluate index after jittering the projection by an angle alpha.}
\usage{
jitterAngle(proj, d, alpha, idx)
}
\arguments{
\item{proj}{Original projection plane}

\item{d}{Data matrix}

\item{alpha}{Jitter angle}

\item{idx}{Index function}
}
\value{
New index value
}
\description{
Re-evaluate index after jittering the projection by an angle alpha.
}
