% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{jitterPoints}
\alias{jitterPoints}
\title{Re-evaluate index after jittering all points by an amount alpha.}
\usage{
jitterPoints(projData, alpha, idx)
}
\arguments{
\item{projData}{Original projected data points}

\item{alpha}{Jitter amount (passed into the jitter() function)}

\item{idx}{Index function}
}
\value{
New index value
}
\description{
Re-evaluate index after jittering all points by an amount alpha.
}
