% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{pipe_data}
\alias{pipe_data}
\title{Generating a sample of points on a pipe}
\usage{
pipe_data(n, p, t = 0.1)
}
\arguments{
\item{n}{number of sample points to generate}

\item{p}{sample dimensionality}

\item{t}{thickness of circle, default=0.1}
}
\value{
sample points in tibble format
}
\description{
Points are drawn from a uniform distribution between -1 and 1,
the pipe structure is generated by rejecting points if they are
not on a circle with radius 1 and thickness t in the last two parameters.
}
\examples{
pipe_data(100, 4)
pipe_data(100, 2, 0.5)
}
