% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingFunctions.R
\name{pipeData}
\alias{pipeData}
\title{Generating a sample of points on a pipe}
\usage{
pipeData(n, p, t = 0.1)
}
\arguments{
\item{n}{sample dimensionality}

\item{p}{number of sample points to generate}

\item{t}{thickness of circle, default=0.1}
}
\value{
sample points in tibble format
}
\description{
Points are drawn from a uniform distribution between -1 and 1,
the pipe structure is generated by rejecting points if they are
not on a circle with radius 1 and thickness t in the last two parameters.
}
\examples{
pipeData(4, 100)
pipeData(2, 100, 0.5)
}
