% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{plot_rotation}
\alias{plot_rotation}
\title{Plot rotation traces of indexes obtained with profileRotation.}
\usage{
plot_rotation(res_mat)
}
\arguments{
\item{res_mat}{data (result of profileRotation)}
}
\value{
ggplot visualisation of the tracing data
}
\description{
Plot rotation traces of indexes obtained with profileRotation.
}
