% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{plot_smoothing_comparison}
\alias{plot_smoothing_comparison}
\title{Plot the comparison of smoothing methods.}
\usage{
plot_smoothing_comparison(sm_mat, lPos = "none")
}
\arguments{
\item{sm_mat}{Result from compare_smoothing}

\item{lPos}{Legend position passed to ggplot2 (default is none for no legend shown)}
}
\value{
ggplot visualisation of the comparison
}
\description{
Plotting method for the results of compareSmoothing.
The results are mapped by facetting over values of alpha and
mapping the method (jitter_angle, jitter_points, no_smoothing) to
linestyle and color (black dashed, black dotted, red solid). By
default legend drawing is turned off, but can be turned on via
the lPos argument, e.g. setting to "bottom" for legend below the plot.
}
