% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot traces of indexes obtained with \code{\link{get_trace}}.}
\usage{
plot_trace(res_mat, rescY = TRUE)
}
\arguments{
\item{res_mat}{data (result of get_trace)}

\item{rescY}{bool to fix y axis range to [0,1]}
}
\value{
ggplot visualisation of the tracing data
}
\description{
Plot traces of indexes obtained with \code{\link{get_trace}}.
}
