% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{plotRotation}
\alias{plotRotation}
\title{Plot rotation traces of indexes obtained with profileRotation.}
\usage{
plotRotation(resMat)
}
\arguments{
\item{resMat}{data (result of profileRotation)}
}
\value{
ggplot visualisation of the tracing data
}
\description{
Plot rotation traces of indexes obtained with profileRotation.
}
