% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_functions.R
\name{scag_index}
\alias{scag_index}
\alias{mine_index}
\alias{mine_indexE}
\alias{holesR}
\alias{cmassR}
\title{Matching index functions to the required format.}
\usage{
scag_index(index_name)

mine_index(index_name)

mine_indexE(index_name)

holesR()

cmassR()
}
\arguments{
\item{index_name}{Index name to select from group of indexes.}
}
\value{
function taking 2-d data matrix and returning the index value
}
\description{
These are convenicence functions that format
scagnostics and mine index functions for direct
use with the guided tour or other functionalities in this package.
}
\section{Functions}{
\itemize{
\item \code{scag_index()}: Scagnostics index from cassowaryr package

\item \code{mine_index()}: MINE index from minerva package

\item \code{mine_indexE()}: MINE index from minerva package (updated estimator)

\item \code{holesR()}: rescaling the tourr holes index

\item \code{cmassR()}: rescaling the tourr cmass index

}}
