% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexFunctions.R
\name{scagIndex}
\alias{scagIndex}
\alias{splineIndex}
\alias{dcorIndex}
\alias{mineIndex}
\alias{mineIndexE}
\alias{holesR}
\alias{cmassR}
\title{Matching index functions to the required format.}
\usage{
scagIndex(indexName)

splineIndex()

dcorIndex()

mineIndex(indexName)

mineIndexE(indexName)

holesR()

cmassR()
}
\arguments{
\item{indexName}{Index name to select from group of indexes.}
}
\value{
function taking 2-d data matrix and returning the index value
}
\description{
These are convenicence functions that format
scagnostics, splines2d, dcor2d and mine index functions for direct
use with the guided tour or other functionalities in this package.
}
\section{Functions}{
\itemize{
\item \code{scagIndex}: Scagnostics index from binostics package

\item \code{splineIndex}: splines2d index from mbgraphic package

\item \code{dcorIndex}: dcor2d index from mbgraphic package

\item \code{mineIndex}: MINE index from minerva package

\item \code{mineIndexE}: MINE index from minerva package (updated estimator)

\item \code{holesR}: rescaling the tourr holes index

\item \code{cmassR}: rescaling the tourr cmass index
}}

