% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{sin_data}
\alias{sin_data}
\title{Generating sine wave sample}
\usage{
sin_data(n, p, f = 1)
}
\arguments{
\item{n}{number of sample points to generate}

\item{p}{sample dimensionality}

\item{f}{jitter factor, default=1}
}
\value{
sample points in tibble format
}
\description{
n-1 points are drawn from a normal distribution with mean=0, sd=1,
the points in the final direction are calculated as the sine of the
values of direction n-1 with additional jittering controled by
the jitter factor f.
}
\examples{
sin_data(100, 4)
sin_data(100, 2, 200)
}
