% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingFunctions.R
\name{sinData}
\alias{sinData}
\title{Generating sine wave sample}
\usage{
sinData(n, p, f = 1)
}
\arguments{
\item{n}{sample dimensionality}

\item{p}{number of sample points to generate}

\item{f}{jitter factor, default=1}
}
\value{
sample points in tibble format
}
\description{
n-1 points are drawn from a normal distribution with mean=0, sd=1,
the points in the final direction are calculated as the sine of the
values of direction n-1 with additional jittering controled by
the jitter factor f.
}
\examples{
sinData(4, 100)
sinData(2, 100, 200)
}
