% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinebil-package.R
\docType{package}
\name{spinebil-package}
\alias{spinebil}
\alias{spinebil-package}
\title{spinebil}
\description{
Functions to evaluate the performance of projection pursuit
index functions using tour methods.
}
\seealso{
The main functions are:
\itemize{
\item \code{\link[=getTrace]{getTrace()}}
\item \code{\link[=profileRotation]{profileRotation()}}
\item \code{\link[=compareSmoothing]{compareSmoothing()}}
\item \code{\link[=timeSequence]{timeSequence()}}
\item \code{\link[=squintAngleEstimate]{squintAngleEstimate()}}
}
}
\author{
\strong{Maintainer}: Ursula Laa \email{ursula.laa@boku.ac.at} (\href{https://orcid.org/0000-0002-0249-6439}{ORCID})

Authors:
\itemize{
  \item Dianne Cook \email{dicook@monash.edu} (\href{https://orcid.org/0000-0002-3813-7155}{ORCID})
}

}
\keyword{internal}
