% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinebil-package.R
\docType{package}
\name{spinebil}
\alias{spinebil}
\alias{spinebil-package}
\title{spinebil}
\description{
Functions to evaluate the performance of projection pursuit
index functions using tour methods.
}
\seealso{
The main functions are:
\itemize{
\item \code{\link[=getTrace]{getTrace()}}
\item \code{\link[=profileRotation]{profileRotation()}}
\item \code{\link[=compareSmoothing]{compareSmoothing()}}
\item \code{\link[=timeSequence]{timeSequence()}}
\item \code{\link[=squintAngleEstimate]{squintAngleEstimate()}}
}
}
