% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{spiral_data}
\alias{spiral_data}
\title{Generating spiral sample}
\usage{
spiral_data(n, p)
}
\arguments{
\item{n}{number of sample points to generate}

\item{p}{sample dimensionality}
}
\value{
sample points in matrix format
}
\description{
n-2 points are drawn from a normal distribution with mean=0, sd=1,
the points in the final two direction are sampled along a spiral
by samping the angle from a normal distribution with mean=0, sd=2*pi
(absolute values are used to fix the orientation of the spiral).
}
\examples{
spiral_data(100, 4)
}
