% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingFunctions.R
\name{spiralData}
\alias{spiralData}
\title{Generating spiral sample}
\usage{
spiralData(n, p)
}
\arguments{
\item{n}{sample dimensionality}

\item{p}{number of sample points to generate}
}
\value{
sample points in matrix format
}
\description{
n-2 points are drawn from a normal distribution with mean=0, sd=1,
the points in the final two direction are sampled along a spiral
by samping the angle from a normal distribution with mean=0, sd=2*pi
(absolute values are used to fix the orientation of the spiral).
}
\examples{
spiralData(4, 100)
}
