% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{distanceDist}
\alias{distanceDist}
\title{Collecting all pairwise distances between input planes.}
\usage{
distanceDist(planes, nn = F)
}
\arguments{
\item{planes}{Input planes (e.g. result of guided tour)}

\item{nn}{Set true to only consider nearest neighbour distances (dummy, need to add this functionality)}
}
\value{
numeric vector containing all distances
}
\description{
The distribution of pairwise distances can be used to understand
differences in space searched with the guided tour approach when
considering different optimisation methods or different number of
noise dimensions.
}
