% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinebil_package.R
\docType{package}
\name{spinebil-package}
\alias{spinebil}
\alias{spinebil-package}
\title{spinebil}
\description{
Functions to evaluate the performance of projection pursuit
index functions using tour methods.
}
\seealso{
The main functions are:
\itemize{
\item \code{\link[=get_trace]{get_trace()}}
\item \code{\link[=profile_rotation]{profile_rotation()}}
\item \code{\link[=compare_smoothing]{compare_smoothing()}}
\item \code{\link[=time_sequence]{time_sequence()}}
\item \code{\link[=squint_angle_estimate]{squint_angle_estimate()}}
}
}
\author{
\strong{Maintainer}: Tina Rashid Jafari \email{tina.rashidjafari@gmail.com} (\href{https://orcid.org/0009-0008-3605-5341}{ORCID})

Authors:
\itemize{
  \item Ursula Laa \email{ursula.laa@boku.ac.at} (\href{https://orcid.org/0000-0002-0249-6439}{ORCID})
  \item Dianne Cook \email{dicook@monash.edu} (\href{https://orcid.org/0000-0002-3813-7155}{ORCID})
}

}
\keyword{internal}
